+++ docs/changes.rst
@@ -15,6 +15,7 @@ Released on March 26th 2024.
 * Avoid crash when installing :ref:`addon-weblate.generate.pseudolocale`.
 * :ref:`azure-setup` gracefully handles repositories with spaces in URL.
 * :ref:`mt-deepl` gracefully handles glossaries for language variants.
+* :doc:`/formats/excel` better handles blank cells.
 
 **Compatibility**